+++ mslicer/src/windows/mod.rs
@@ -15,7 +15,6 @@ mod remote_print;
 mod slice_config;
 mod slice_operation;
 mod supports;
-mod tasks;
 pub mod tools;
 mod top_bar;
 mod workspace;
@@ -33,20 +32,18 @@ pub enum Tab {
     SliceConfig,
     SliceOperation,
     Supports,
-    Tasks,
     Viewport,
     Workspace,
 }
 
 impl Tab {
-    const ALL: [Tab; 8] = [
+    const ALL: [Tab; 7] = [
         Tab::Logs,
         Tab::Models,
         Tab::RemotePrint,
         Tab::SliceConfig,
         Tab::SliceOperation,
         Tab::Supports,
-        Tab::Tasks,
         Tab::Workspace,
     ];
 
@@ -58,7 +55,6 @@ impl Tab {
             Tab::SliceConfig => "Slice Config",
             Tab::SliceOperation => "Slice Operation",
             Tab::Supports => "Supports",
-            Tab::Tasks => "Tasks",
             Tab::Viewport => "Viewport",
             Tab::Workspace => "Workspace",
         }
@@ -80,7 +76,6 @@ impl TabViewer for Tabs<'_> {
             Tab::SliceConfig => slice_config::ui(self.app, ui, self.ctx),
             Tab::SliceOperation => slice_operation::ui(self.app, ui, self.ctx),
             Tab::Supports => supports::ui(self.app, ui, self.ctx),
-            Tab::Tasks => tasks::ui(self.app, ui, self.ctx),
             Tab::Viewport => viewport(self.app, ui, self.ctx),
             Tab::Workspace => workspace::ui(self.app, ui, self.ctx),
         }